Move job retry and backup job cleanup into worker
Previously the job retry was being done in stats, while backup removal was in Manager.  This was a
little scattered, also an issue if stats is falling behind and jobs are removed faster from backup than were retried.
This was also bottlenecked since manager and stats were single processes.
Moved to this logic to Worker - seems to get a little more concurrency as well since it's a process per job.