Don’t add exq supervisor inside phoenix
Reason:

https://hexdocs.pm/exrm/extra-common-issues.html

If you are running into issues with your dependencies missing their
dependencies, it’s likely that the author did not put the dependencies
in the :application block of their mix.exs. You may have to fork, or
issue a pull request in order to resolve this issue. Alternatively, if
you know what the dependency is, you can put it in your own mix.exs,
and the release process will ensure that it is loaded with everything
else.