Flush stats on interval using Redis batch API
Stats updates are currently being sent to Redis one at a time. This causes a situation
where stats updates can fall behind other parts of the system and cause memory to be eaten up.
This commit changes stats to flush batches of updates at a time. This allows the stats to
keep up with other parts of the system and update Redis more efficiently.