Update dep in mix.exs
the UUID lib has changed it's name to `elixir_uuid`. The API is exactly the same. Having the old `uuid` and `elixir_uuid` deps is causing dep clash issues in that they both occupy the same namespace. This PR will bring it in line with the newest version and prevent future namespace clashes.