handle timeouts on middleware pipeline
Currently, if any of the middleware fails, (notably
Exq.Middleware.Job), it prevents the execution of remaining
middlewares. This is highly problematic as Exq.Middleware.Manager is
essential for Exq to function properly.

In this specific case, if the redis operation in Exq.Middleware.Job
timeouts (or fails), then Exq fails to decrement the busy worker count
and with enough number of such failures, the worker eventually will
stop picking up new jobs.

This commit tries to improve the conditions by adding a retry for
`remove_job_from_backup` operation. Retrying is ok in this case as the
operation is idempotent